+++ CHANGELOG.md
@@ -5,13 +5,15 @@ All notable changes to the "direnv" extension will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+
+## [0.5.0] - 2021-12-28
 ### Fixed
 - Show error when direnv executable is missing
 ### Added
 - Make path to direnv executable configurable
 - Offer to install direnv or configure the executable path when it's not found
 
-## [0.4.0]
+## [0.4.0] - 2021-12-12
 ### Fixed
 - Handle null values exported by direnv
 - Creating a new .envrc uses the intended file name
@@ -46,7 +48,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.0.1] - 2021-11-23
 - Initial release
 
-[Unreleased]: https://github.com/direnv/direnv-vscode/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/direnv/direnv-vscode/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/direnv/direnv-vscode/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/direnv/direnv-vscode/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/direnv/direnv-vscode/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/direnv/direnv-vscode/compare/v0.1.0...v0.2.0